+++ src/database/models/__tests__/fixtures/embedding.ts
