+++ src/database/models/__tests__/sessionGroup.test.ts
@@ -1,13 +1,14 @@
 // @vitest-environment node
 import { eq } from 'drizzle-orm/expressions';
-import { afterEach, beforeEach, describe, expect, it } from 'vitest';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { getTestDBInstance } from '@/database/server/core/dbForTest';
+import { LobeChatDatabase } from '@/database/type';
 
-import { sessionGroups, users } from '../../../schemas';
-import { SessionGroupModel } from '../sessionGroup';
+import { sessionGroups, users } from '../../schemas';
+import { SessionGroupModel } from '../../server/models/sessionGroup';
+import { getTestDB } from './_util';
 
-let serverDB = await getTestDBInstance();
+const serverDB: LobeChatDatabase = await getTestDB();
 
 const userId = 'session-group-model-test-user-id';
 const sessionGroupModel = new SessionGroupModel(serverDB, userId);
