+++ src/database/models/__tests__/topic.test.ts
@@ -3,8 +3,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { LobeChatDatabase } from '@/database/type';
 
-import { messages, sessions, topics, users } from '../../../schemas';
-import { CreateTopicParams, TopicModel } from '../topic';
+import { messages, sessions, topics, users } from '../../schemas';
+import { CreateTopicParams, TopicModel } from '../../server/models/topic';
 import { getTestDB } from './_util';
 
 const serverDB: LobeChatDatabase = await getTestDB();
