+++ vitest.config.ts
@@ -30,6 +30,7 @@ export default defineConfig({
       '**/node_modules/**',
       '**/dist/**',
       '**/build/**',
+      'src/database/server/**/**',
       'src/database/repositories/dataImporter/**/**',
     ],
     globals: true,
